pyproject.toml
[tool.poetry]
name = "risk-module-server"
version = "0.1.0"
description = ""
authors = ["Togglecorp Dev dev@togglecorp.com"]
license = "GNU Affero General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.8"
celery = { version = "==5.1.1", extras = ["redis"] }
djangorestframework-camel-case = "==1.2.0"
djangorestframework = "==3.12.4"
django-celery-beat = "==2.2.0"
django-cors-headers = "==3.7.0"
django-crispy-forms = "==1.12.0"
django-storages = "==1.11.1"
django-enumfield = "==2.0.2"
django-filter = "==2.4.0"
django-redis = "==5.0.0"
Django = ">=3.2,<3.3"
ipython = "*"
factory-boy = "*"
drf-writable-nested = "*"
openpyxl = "==3.0.9"
psycopg2-binary = "==2.9.1"
requests = "*"
geopy = "==2.2.0"
gunicorn = "~=20.0.4"

[tool.poetry.dev-dependencies]
pytest-django = "*"

[tool.pyright]
extraPaths = ["apps"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
]
reportMissingImports = true
reportMissingTypeStubs = false

[tool.black]
line-length = 125
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
)/
'''
extend-exclude = '^.*\b(migrations)\b.*$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"